backend/requirements.txt
# CCS Quote Tool v2 - Backend Dependencies
# FastAPI + Multi-tenant SaaS Platform

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Redis & Caching
redis==4.6.0
aioredis==2.0.1

# Background Tasks
celery==5.3.4
celery[redis]==5.3.4

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# AI & OpenAI
openai==1.3.8

# Data Processing
beautifulsoup4==4.12.2
lxml==5.1.0
python-dateutil==2.8.2

# Email
fastapi-mail==1.4.1
jinja2==3.1.2

# File Processing
python-magic==0.4.27
aiofiles==23.2.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
psycopg2-binary==2.9.9

# Additional Utilities
click==8.1.7
python-dotenv==1.0.0
email-validator==2.1.0
phonenumbers==8.13.26
